API/requirements.txt
matplotlib
numpy
opencv_python
Pillow
streamlit
fastapi
uvicorn
python-multipart
